[fix] class MinimumCurvature - method compute
distinguishes as follows:
1. in case given strain is within the minimal positive and negative maximum strains, then edge-strains with minimum absolute curvature are decisive
2. in case given strain is outside the minimal positive and negative maximum strains, then edge-strains with maximum absolute curvature are decisive

Background #1:
--------------
The minimum absolute curvature is needed to get the absolute minimum.
Furthermore, the computed minimum curvatures are more or less a assumption by the author.
Assuming that these values fit 99.9 % of cases.

Background #2:
--------------
The maximum absolute curvature is needed as otherwise the strain would exceed in some material-curves